requirement.txt
Flask==3.0.0
groq==0.1.0
numpy==1.24.3
sentence-transformers==2.2.2
scikit-learn==1.3.2
python-dotenv==1.0.0
httpx[http2]==0.26.0
googletrans==4.0.0-rc1
fuzzywuzzy==0.18.0